move CheckPasswordConfirm under security from cmd
